// src/pages/CIFInsert/Review/types.ts
export interface CIFReviewFormData {
  title: string;
  lastName: string;
  firstName: string;
  middleName: string;
  suffix: string;
  gender: string;
  maritalStatus: string;
  birthDate: string;
  age: number;
  birthPlace: string;
  nationality: string;
  previousName: string;
  nickName: string;
  bloodType: string;
  weight: number;
  height: number;
  isDeceased: boolean;
  isBeneficiary: boolean;
  primaryBeneficiary: string;
  education: string;
  schoolLevel: string;
  elementaryDependents: number;
  highSchoolDependents: number;
  collegeDependents: number;
  noOfDependents: number;
  numberOfCarsOwned: number;
  companyType: string;
  companyName: string;
  registeredName: string;
  tin: string;
  startOfBusiness: string;
  numberOfEmployees: number;
  trader1: string;
  trader2: string;
  firmSize: string;
  conglomerateDomain: string;
  entityLocation: string;
  countryOfOrigin: string;
  placeOfRegistration: string;
  legalForm: string;
  companyIndustry: string;
  originOfEntity: string;
  companyBusinessActivity: string;
  netTaxableIncome: number;
  monthlyExpenses: number;
  parentClient: string;
  roleOfParent: string;
  contactPerson: string;
  designation: string;
  organizationRemarks: string;
  organizationBusinessActivity: string;
  organizationNationality: string;
  address1Used: string;
  address2Used: string;
  primaryID1: string;
  primaryID2: string;
  primaryID3: string;
  secondaryID1: string;
  secondaryID2: string;
  secondaryID3: string;
  primaryContact: string;
  secondaryContact: string;
  presentedIDType: string;
  idNoPresentedForAMLA: string;
  countryCode: string;
  mobileNumber: string;
  emailAddress: string;
  insufficientInformation: boolean;
  insufficientInformationRemarks: string;
  insufficientDocuments: boolean;
  insufficientDocumentsRemarks: string;
  employeeID: string;
  occupation: string;
  status: string;
  position: string;
  inclusiveDate: string;
  startDate: string;
  endDate: string;
  lengthOfService: number;
  companyEmployerName: string;
  employerTIN: string;
  employerAddress: string;
  employerContactPerson: string;
  employerContactNo: string;
  businessActivities: string;
  industry: string;
  businessName: string;
  businessActivity: string;
  businessMainAddress: string;
  businessAdditionalAddress: string;
  businessID1: string;
  businessID2: string;
  businessPrimaryContact: string;
  businessSecondaryContact: string;
  otherBusinessName: string;
  otherBusinessActivity: string;
  otherBusinessMainAddress: string;
  otherBusinessAdditionalAddress: string;
  otherBusinessID1: string;
  otherBusinessID2: string;
  otherBusinessPrimaryContact: string;
  otherBusinessSecondaryContact: string;
  salaryIndicator: string;
  fundSource: string;
  salary: number;
  grossIncome: number;
  otherIncome: number;
  otherIncomeSource: string;
  otherIncomeSourceAmount: number;
  monthlyAverageIncome: number;
  isLargeExposure: boolean;
  isMicrofinanceBorrower: boolean;
  isRegularLoanBorrower: boolean;
  isComaker: boolean;
  isSavingsAcctDepositor: boolean;
  isCurrentAcctDepositor: boolean;
  isTimeDepositDepositor: boolean;
  isSpecialSavingsDepositor: boolean;
  enforceCreditLimit: boolean;
  originalBalance: number;
  outstandingBalance: number;
  pesonet: number;
  instapay: number;
  mobileWallet: number;
  isBankEmployeeRelated: boolean;
  bankEmployeeName: string;
  relationship: string;
  dosri: string;
  isBankEmployee: boolean;
  employeeType: string;
  isPEP: boolean;
  pepPosition: string;
  pepPlace: string;
  pepTerm: string;
  isWatchListed: boolean;
  isLinkedAccount: boolean;
  isPayee: boolean;
  relatedParty: string;
  overallScore: number;
  classification: string;
  customerDueDiligence: string;
  group1: string;
  group2: string;
  group3: string;
  customUse1: string;
  customUse2: string;
  customUse3: string;
  customUse4: string;
  remarks: string;
  clientStatus: string;
  memberSinceDate: string;
  lastClientUpdate: string;
  clientPicture1: string;
  clientPicture2: string;
  signaturePicture1: string;
  signaturePicture2: string;
}

export interface CIFReviewSelectedTypes {
  individual: boolean;
  company: boolean;
  government: boolean;
  organization: boolean;
}

export interface CIFReviewState {
  formData: CIFReviewFormData;
  selectedTypes: CIFReviewSelectedTypes;
}